arm64: implement bounded memory64 execution
Why:
- ARM64 still rejected memory64 despite having part of the bulk-operation width plumbing

What:
- decode full-u64 scalar, float, and SIMD memarg offsets
- add carry-checked full-width effective-address and dynamic bulk bounds checks
- preserve memory64 widths for size, grow, copy, fill, and data initialization
- reject wrapping addresses and wide grow deltas with canonical i64 failure values
- make dead-code immediate walking memory-width aware
- admit bounded single-memory memory64 on Linux and Darwin ARM64 explicit bounds
- add native scalar, growth, wraparound-trap, and codec coverage

Proof:
- go test ./...
- go test -race ./src/core/compiler/wasm ./src/wago
- go test -tags wago_guardpage ./src/wago
- Linux/ARM64 and Darwin/ARM64 wago test-binary cross-compilation
- git diff --check

+++ src/core/compiler/backend/railshot/arm64/compile.go
@@ -357,7 +357,7 @@ type storeForward struct {
 	reg    Reg
 	typ    machineType
 	local  int
-	offset uint32
+	offset uint64
 	size   int
 }
